Separating different tasks into separate classes.
